backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioice==0.10.1
aiortc==1.11.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
av==14.3.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.33.0
azure-identity==1.21.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudevents==1.11.0
cryptography==44.0.2
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.12
frozenlist==1.6.0
google-crc32c==1.7.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.6.1
isodate==0.7.2
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.11.0
MarkupSafe==3.0.2
more-itertools==10.7.0
msal==1.32.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.4.3
nest-asyncio==1.6.0
numpy==2.2.5
oauthlib==3.2.2
openai==1.76.0
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
packaging==25.0
parse==1.20.2
pathable==0.4.4
prance==23.6.21.0
propcache==0.3.1
pybars4==0.9.13
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pyee==13.0.0
PyJWT==2.10.1
pylibsrtp==0.12.0
PyMeta3==0.5.1
pyOpenSSL==25.0.0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rpds-py==0.24.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scipy==1.15.2
semantic-kernel==1.28.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
Werkzeug==3.1.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
